requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
facenet-pytorch==2.6.0
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.11.0.86
packaging==26.0
pillow==10.2.0
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
psycopg2-binary==2.9.11
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
PyYAML==6.0.3
requests==2.32.5
scipy==1.15.3
six==1.17.0
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.3
typing_extensions==4.15.0
ultralytics==8.4.12
ultralytics-thop==2.0.18
urllib3==2.6.3
